refactor: Extract data selectors
First step before using selectors directly inside mapStateToProps
End goal is

- remove data mangling from rendering
- have memoization on the selectors
- be able to have the chart reflect the month currently selected
